requirements.txt
# ============================================
# AutoHire AI - Automated Job Application Tracker
# Requirements File
# ============================================

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
mysql-connector-python==8.2.0
SQLAlchemy==2.0.23

# Web Scraping & Automation
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0

# Scheduling
APScheduler==3.10.4

# Email
secure-smtplib==0.1.1

# Environment Variables
python-dotenv==1.0.0

# Utilities
Jinja2==3.1.2
